fakejob-main/requirements.txt
Flask==3.0.3
Flask-Cors==5.0.0
gunicorn==23.0.0
tensorflow-cpu==2.15.1
numpy==1.26.4
lime==0.2.0.1
PyPDF2==3.0.1
python-docx==1.1.2
pytesseract==0.3.13
Pillow==10.4.0
